# backend/docs/swagger.yaml
basePath: /api
definitions:
  handlers.Response:
    properties:
      message:
        type: string
    type: object
  inputs.Airport:
    properties:
      address:
        type: string
      latitude:
        type: number
      longitude:
        type: number
      name:
        type: string
    type: object
  inputs.CreateFlight:
    properties:
      arrival:
        $ref: '#/definitions/inputs.Airport'
      departure:
        $ref: '#/definitions/inputs.Airport'
    required:
    - arrival
    - departure
    type: object
...
  responses.ListUser:
    properties:
      created_at:
        type: string
      email:
        type: string
      first_name:
        type: string
      id:
        type: integer
      is_verified:
        type: boolean
      last_name:
        type: string
      role:
        type: string
      updated_at:
        type: string
    type: object
...
  responses.ResponseAirport:
    properties:
      address:
        type: string
      latitude:
        type: number
      longitude:
        type: number
      name:
        type: string
    type: object
  responses.ResponseFlight:
    properties:
      arrival:
        $ref: '#/definitions/responses.ResponseAirport'
      created_at:
        type: string
      departure:
        $ref: '#/definitions/responses.ResponseAirport'
      id:
        type: integer
      pilot:
        $ref: '#/definitions/responses.ListUser'
      pilot_id:
        type: integer
      price:
        type: number
      status:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
      vehicle:
        $ref: '#/definitions/responses.Vehicle'
      vehicle_id:
        type: integer
    required:
    - arrival
    - departure
    - id
    - status
    type: object
...
  responses.Vehicle:
    properties:
      created_at:
        type: string
      cruise_altitude:
        type: number
      cruise_speed:
        type: number
      id:
        type: integer
      is_selected:
        type: boolean
      is_verified:
        type: boolean
      matriculation:
        type: string
      model_name:
        type: string
      seat:
        type: integer
      type:
        type: string
      updated_at:
        type: string
    type: object
info:
  contact: {}
  description: this is the go project.
  title: Backend AirFleet
  version: "1.0"
paths:
  /flights:
    get:
      consumes:
      - application/json
      description: get all flights
      parameters:
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/responses.ResponseFlight'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.Response'
      security:
      - BearerAuth: []
      summary: Get all flights
      tags:
      - flight
    post:
      consumes:
      - application/json
      description: create a new flight
      parameters:
      - description: Message body
        in: body
        name: flightInput
        required: true
        schema:
          $ref: '#/definitions/inputs.CreateFlight'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/responses.ResponseFlight'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.Response'
      security:
      - BearerAuth: []
      summary: Create flight
      tags:
      - flight
  /flights/{id}:
    get:
      consumes:
      - application/json
      description: get flight
      parameters:
      - description: Flight ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.ResponseFlight'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.Response'
      security:
      - BearerAuth: []
      summary: Get flight
      tags:
      - flight
  /flights/current:
    get:
      consumes:
      - application/json
      description: get current flight
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.ResponseFlight'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.Response'
      security:
      - BearerAuth: []
      summary: Get current flight
      tags:
      - flight
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"